mapping with diamond and plastic

+++ macro/beamtime/cern2016/init_calib.sh
@@ -24,8 +24,12 @@ cp rootlogon.C ${cRun}
 cp .rootrc ${cRun}
 cd ${cRun}
 
+# Global variables, for for-loops. Runs from 1-15 and 16-79 in two for-loops
+iRestart=0
 iStep=0
 iStepLast=0
+
+# ************************** Starting while Loop ***************************** #
 while [[ $nIter > 0 ]]; do 
 echo start from scratch with $nIter iterations 
 if((0)); then
@@ -55,16 +59,18 @@ if [[ ${cSel} = "-" ]];then
     cSel="0"
 fi
 #copy calibration file 
-#cp -v tofTestBeamClust_${cRun}.hst.root ${cRun}_${cMode}_${cSel}tofTestBeamClust.hst.root
-cp -v ../Init${iStepLast}/tofTestBeamClust_${cRun}_set${iCalSet}.hst.root ${cRun}_set${iCalSet}_${cMode}_${cSel}tofTestBeamClust.hst.root
+  if (($iStep > $iRestart)) ; then 
+    cp -v ../Init${iStepLast}/tofTestBeamClust_${cRun}_set${iCalSet}.hst.root ${cRun}_set${iCalSet}_${cMode}_${cSel}tofTestBeamClust.hst.root
+  fi
 fi 
-
-# generate new calibration file
-root -b -q '../../ana_digi_ini.C('$inOpt',"'${cRun}'",'${iCalSet}','${iBRef}') '
 lastOpt=$inOpt
 
-#./screenshot.sh
-cp *pdf ../
+if (($iStep > $iRestart)) ; then
+  # generate new calibration file
+  root -b -q '../../ana_digi_ini.C('$inOpt',"'${cRun}'",'${iCalSet}','${iBRef}') '
+  #./screenshot.sh
+  cp *pdf ../
+fi
 cd .. 
 done
 
@@ -98,21 +104,27 @@ if [[ ${cSel} = "-" ]];then
     cSel="0"
 fi
 #copy calibration file 
-cp -v ../Init${iStepLast}/tofTestBeamClust_${cRun}_set${iCalSet}.hst.root ${cRun}_set${iCalSet}_${cMode}_${cSel}tofTestBeamClust.hst.root
+if (($iStep > $iRestart)) ; then
+  cp -v ../Init${iStepLast}/tofTestBeamClust_${cRun}_set${iCalSet}.hst.root ${cRun}_set${iCalSet}_${cMode}_${cSel}tofTestBeamClust.hst.root
+fi
 fi 
 
-# generate new calibration file
-root -b -q '../../ana_digi_cal.C('$inOpt',"'${cRun}'",'${iCalSet}','${iBRef}') '
-
 lastOpt=$inOpt
+# generate new calibration file
+if (($iStep > $iRestart)) ; then 
+  root -b -q '../../ana_digi_cal.C('$inOpt',"'${cRun}'",'${iCalSet}','${iBRef}') '
 
-cp -v tofTestBeamClust_${cRun}_set${iCalSet}.hst.root ../${cRun}_set${iCalSet}_${cMode}_${cSel}tofTestBeamClust.hst.root
-cp *pdf ../
+  cp -v tofTestBeamClust_${cRun}_set${iCalSet}.hst.root ../${cRun}_set${iCalSet}_${cMode}_${cSel}tofTestBeamClust.hst.root
+  cp *pdf ../
 #./screenshot.sh
-cd .. 
-rm ../${cRun}_set${iCalSet}_${cMode}_${cSel}tofTestBeamClust.hst.root
-ln -s ./${cRun}/${cRun}_set${iCalSet}_${cMode}_${cSel}tofTestBeamClust.hst.root ../${cRun}_set${iCalSet}_${cMode}_${cSel}tofTestBeamClust.hst.root
-
+  cd .. 
+  rm ../${cRun}_set${iCalSet}_${cMode}_${cSel}tofTestBeamClust.hst.root
+  ln -s ./${cRun}/${cRun}_set${iCalSet}_${cMode}_${cSel}tofTestBeamClust.hst.root ../${cRun}_set${iCalSet}_${cMode}_${cSel}tofTestBeamClust.hst.root
+echo Init step $iStep with mode ${cMode}, option $inOpt  finished
+else 
+  cd ..
+  echo Init step $iStep with mode ${cMode}, option $inOpt  skipped
+fi
 done
 (( nIter -= 1))
 done
